requirements.txt
build
faiss-cpu
numpy
openai
python-frontmatter
sentence-transformers
tqdm
twine
